+++ src/main/java/com/thoughtworks/com/mapper/ProductWithPrice.java
@@ -2,17 +2,10 @@ package com.thoughtworks.com.mapper;
 
 import java.util.Date;
 
-/**
- * Created with IntelliJ IDEA.
- * User: twer
- * Date: 6/20/14
- * Time: 6:37 PM
- * To change this template use File | Settings | File Templates.
- */
 public class ProductWithPrice {
-    public long id;
+    public int id;
     public String name;
-    public long productId;
+    public int productId;
     public Date time;
     public double price;
 }
